test_agent/requirements.txt
google-adk>=1.22.0
a2a-sdk>=0.3.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
